requirements.txt
aiohttp==3.5.4
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==19.1.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
discord==1.0.1
discord.py==1.2.3
future==0.17.1
idna==2.8
idna-ssl==1.1.0
multidict==4.5.2
pycparser==2.19
python-telegram-bot==11.1.0
six==1.12.0
typing-extensions==3.7.4
websockets==6.0
yarl==1.3.0
